feat: Improve existing stepper ui.
