+++ package.json
@@ -26,6 +26,7 @@
     "@typescript-eslint/eslint-plugin": "^5.41.0",
     "apexcharts": "^3.36.0",
     "axios": "^1.1.3",
+    "chakra-ui-steps": "^1.8.0",
     "framer-motion": "^7.6.2",
     "mongodb": "^4.11.0",
     "mysql2": "^2.3.3",
