--- src/contexts/store/constants.js
@@ -1,3 +1,0 @@
-export const SET_STEPS = 'SET_STEPS'
-export const INCREMENT_CURRENT_STEP = 'INCREMENT_CURRENT_STEP'
-export const DECREMENT_CURRENT_STEP = 'DECREMENT_CURRENT_STEP'
